feat: add next plugin that deals with webpack workarounds
The canvas package, which is a dependency of jsdom, has a native binding which causes the following error during Next.js build process: "Module did not self-register".
We circuvent that by using a null-loader.
See https://github.com/zeit/next.js/issues/7894

The ws package, which is also a dependency of jsdom, tries to optionally load some dependencies.
This produces a warning in webpack that we want to avoid.
We circuvent that by adding them to externals.